CRM-呼叫中心-呼叫功能抽像出工具方法
